test: move and expand hidden quit commands test
